Rafactor listing models and listing controller
