+++ server/src/models/listing/listing.model.ts
@@ -1,6 +1,7 @@
-import { Schema } from 'mongoose';
+import { Schema, Model } from 'mongoose';
 
 const mongoose = require('mongoose');
+const { Pet } = require('./pet/pet.model');
 
 const itemTypes = ['Pet']; // Add more types here: 'Service', 'Product', etc.
 const saleTypes = ['Adoption', 'Sale']; // Add more types here: 'Subscriptions', 'Rentals', etc.
@@ -8,25 +9,47 @@ const mediaTypes = ['image']; // Add more types here: 'video', etc.
 
 export interface IListing {
   _id: Schema.Types.ObjectId;
-  listerId: Schema.Types.ObjectId;
-  itemId: Schema.Types.ObjectId;
+  lister: Schema.Types.ObjectId;
+  item: Schema.Types.ObjectId;
   price: number;
   description: string;
   itemType: string;
   saleType: string;
-  media: { type: string, url: string }[];
+  media: { type: string; url: string }[];
   createdAt: Date;
   updatedAt: Date;
 }
-  
-const ListingSchema = new Schema(
+
+const listingSchema = new Schema(
   {
-    listerId: { type: Schema.Types.ObjectId, immutable: true, required: true },
-    itemId: { type: Schema.Types.ObjectId, immutable: true, refPath: 'itemType', required: true },
+    lister: {
+      type: Schema.Types.ObjectId,
+      immutable: true,
+      ref: 'User',
+      required: true,
+      autopopulate: true
+    },
+    item: {
+      type: Schema.Types.ObjectId,
+      immutable: true,
+      refPath: 'itemType',
+      required: true,
+      autopopulate: true,
+    },
     price: { type: Number, required: true },
     description: { type: String, required: true },
-    itemType: { type: String, enum: itemTypes, immutable: true, required: true },
-    saleType: { type: String, enum: saleTypes, immutable: true, required: true },
+    itemType: {
+      type: String,
+      enum: itemTypes,
+      immutable: true,
+      required: true
+    },
+    saleType: {
+      type: String,
+      enum: saleTypes,
+      immutable: true,
+      required: true
+    },
     createdAt: { type: Date, immutable: true, default: () => Date.now() },
     updatedAt: { type: Date, default: () => Date.now() },
     media: [{ type: { type: String, enum: mediaTypes }, url: String }]
@@ -34,7 +57,8 @@ const ListingSchema = new Schema(
   { collection: 'listings' }
 );
 
-const Listing = mongoose.model('Listing', ListingSchema);
+listingSchema.plugin(require('mongoose-autopopulate'));
+
+const Listing = mongoose.model('Listing', listingSchema);
 
 module.exports = { Listing, itemTypes, saleTypes, mediaTypes };
-
