+++ server/src/models/listing/pet/animal/dog/dog.model.ts
@@ -7,34 +7,36 @@ const hairCoats = ["Double", "Silky", "Wire", "Curly", "Hairless", "Long", "Medi
 
 export interface IDog {
   _id: Schema.Types.ObjectId;
-  breedIds: Schema.Types.ObjectId[];
-  vaccineIds: Schema.Types.ObjectId[];
+  breeds: Schema.Types.ObjectId[];
+  vaccines: Schema.Types.ObjectId[];
   size: string;
   weight: number;
   hairCoat: string;
   isHypoallergenic: boolean;
   isMicrochipped: boolean;
   isNeutered: boolean;
-  isPottyTrained: boolean;
   isHdbApproved: boolean;
+  avsLicenseNumber: string;
 }
 
-const DogSchema = new Schema(
+const dogSchema = new Schema(
   {
-    breedIds: [{ type: Schema.Types.ObjectId, immutable: true, ref: 'DogBreed', required: true }],
-    vaccineIds: [{ type: Schema.Types.ObjectId, ref: 'Vaccine', required: true }],
+    breeds: [{ type: Schema.Types.ObjectId, immutable: true, ref: 'DogBreed', required: true, autopopulate: true }],
+    vaccines: [{ type: Schema.Types.ObjectId, ref: 'DogVaccine', required: true, autopopulate: true }],
     size: { type: String, enum: sizes, required: true },
     weight: { type: Number, required: true },
     hairCoat: { type: String, enum: hairCoats, required: true },
     isHypoallergenic: { type: Boolean, required: true },
     isMicrochipped: { type: Boolean, required: true },
     isNeutered: { type: Boolean, required: true },
-    isPottyTrained: { type: Boolean, required: true },
-    isHdbApproved: { type: Boolean, required: true }
+    isHdbApproved: { type: Boolean, required: true },
+    avsLicenseNumber: { type: String, required: true }
   },
   { collection: 'dogs' }
 );
 
-const Dog = mongoose.model('Dog', DogSchema);
+dogSchema.plugin(require('mongoose-autopopulate'));
+
+const Dog = mongoose.model('Dog', dogSchema);
 
 module.exports = { Dog, sizes, hairCoats };
